Distro: move `rhel7` -> `rhel/rhel7`
