+++ pkg/distro/rhel/rhel7/azure.go
