+++ pkg/distro/rhel/rhel7/images.go
