+++ pkg/distro/rhel/rhel7/imagetype.go
