+++ pkg/distro/rhel/rhel7/package_sets.go
