+++ pkg/distro/rhel/rhel7/partition_tables.go
