+++ pkg/distro/rhel/rhel7/qcow2.go
